requirements.txt
# Core dependencies
pandas>=2.2.0
python-dotenv>=0.22.0
tqdm>=4.66.1

# Vector database and embeddings
chromadb>=0.4.22
sentence-transformers>=2.2.2

# LLM APIs
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.4.0

# Async and HTTP
aiohttp>=3.9.1
asyncio>=3.4.3

# Optional: Data visualization (only if you plan to use these features)
# jupyter>=1.0.0
# matplotlib>=3.8.2
# plotly>=5.18.0